% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{apply_mask}
\alias{apply_mask}
\title{Removes all masked entries from the array and returns the result}
\usage{
apply_mask(x)
}
\arguments{
\item{x}{A \code{MaskedArray} object}
}
\value{
The same type as \code{x@value}, but shorter
}
\description{
Removes all masked entries from the array and returns the result
}
