% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked_vector.R
\name{combine_masks}
\alias{combine_masks}
\title{Returns a new object whose mask is the union of the existing mask, and a new mask}
\usage{
combine_masks(x, indices)
}
\arguments{
\item{indices}{A list of index vectors}
}
\description{
Returns a new object whose mask is the union of the existing mask, and a new mask
}
