% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Create a SHA1 dictionary from a dataframe column}
\usage{
dictionary(.data, .var)
}
\arguments{
\item{.data}{Dataframe containing the variable}

\item{.var}{Column to be used to create the dictionary}
}
\description{
Given a dataframe variable, returns a dictionary to be used to crypt and
decrypt this variable using \code{crypt()} and \code{decrypt()} functions
}
