% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionnary.R
\name{dictionnary}
\alias{dictionnary}
\title{Create a SHA1 dictionnary from a dataframe column}
\usage{
dictionnary(.data, .var)
}
\arguments{
\item{.data}{Dataframe containing the variable}

\item{.var}{Column to be used to create the dictionnary}
}
\description{
Given a dataframe variable, returns a dictionnary to be used to crypt and
decrypt this variable using \code{crypt()} and \code{decrypt()} fuctions
}

