% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{data}
\name{experiment}
\alias{experiment}
\title{Fake experiment.}
\format{A data frame with three variables: \code{participant},
  \code{condition}, and \code{score}}
\usage{
experiment
}
\description{
A fake experiment dataset.
}
\keyword{datasets}
