% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked_vector.R
\name{invert_index}
\alias{invert_index}
\title{Given a vector that will be used as an index,
returns a vector of the same type that returns
the inverse of the original vector when used as an index}
\usage{
invert_index(index)
}
\description{
Given a vector that will be used as an index,
returns a vector of the same type that returns
the inverse of the original vector when used as an index
}
