% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked_vector.R
\docType{class}
\name{MaskedArray-class}
\alias{MaskedArray-class}
\title{The mask is a vector that will be used to filter the inner vector}
\description{
The mask is a vector that will be used to filter the inner vector
}
