% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked_vector.R
\name{MaskedArray}
\alias{MaskedArray}
\title{MaskedArray constructor function}
\usage{
MaskedArray(value, mask = NULL, masks = NULL)
}
\arguments{
\item{value}{An array-like object (vector, matrix, array, data.frame etc)
to mask}

\item{mask}{A vector}
}
\description{
MaskedArray constructor function
}
