% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize}
\alias{randomize}
\title{Randomize rows' order}
\usage{
randomize(.data)
}
\arguments{
\item{.data}{Dataframe containing rows' order to be randomized}
}
\description{
Randomize rows' order of a give data frame
}

