% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reset_mask}
\alias{reset_mask}
\title{Resets the mask of a \code{MaskedArray}, returning a new \code{MaskedArray} which
has no masked/hidden elements}
\usage{
reset_mask(x)
}
\arguments{
\item{x}{A \code{MaskedArray} object}
}
\value{
A \code{MaskedArray} object
}
\description{
Resets the mask of a \code{MaskedArray}, returning a new \code{MaskedArray} which
has no masked/hidden elements
}
