% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{unmask}
\alias{unmask}
\title{Removes the mask of a \code{MaskedArray}, exposing the full array}
\usage{
unmask(x)
}
\arguments{
\item{x}{A \code{MaskedArray} object}
}
\value{
The same type as \code{x@value}: some array-like object.
}
\description{
Removes the mask of a \code{MaskedArray}, exposing the full array
}
