% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visnonp-package.R
\docType{package}
\name{visnonp-package}
\alias{visnonp}
\alias{visnonp-package}
\title{visnonp: Visual Class for Vectors with Non-Publishing Requirements}
\description{
Create vectors with sticky flags for elements that should not be displayed. Numeric vectors have basic subset and arithmetic methods implemented.
}
\author{
\strong{Maintainer}: Ian Powell \email{powell.ian.n@gmail.com}

}
\keyword{internal}
