% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masked_vector.R
\name{apply_mask}
\alias{apply_mask}
\title{Removes all masked entries from the array and returns the result}
\usage{
apply_mask(x)
}
\description{
Removes all masked entries from the array and returns the result
}
