% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt one column from a crypted dataframe using a dictionary}
\usage{
decrypt(.data, .var, dictionary)
}
\arguments{
\item{.data}{Dataframe containing crypted variable}

\item{.var}{Column to be decrypted}

\item{dictionary}{Dictionary to be used to decrypt the crypted variable}
}
\description{
Given a maskr crypted dataframe and a dictionary, returns a decrypted data
frame.
}
