% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt multiple columns from a crypted dataframe}
\usage{
decrypt(.data, .dic)
}
\arguments{
\item{.data}{Dataframe containing crypted variable}

\item{.dic}{Dictionary containing the cryptograms}
}
\description{
Given a maskr crypted dataframe and some column variables, returns a decrypted
data frame.
}
\examples{
data(mtcars)
dic <- dictionary(mtcars, cyl, vs)
mtcars_encrypted <- encrypt(mtcars, dic)
mtcars_decrypted <- decrypt(mtcars_encrypted, dic)

}
