% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt one column from a crypted dataframe}
\usage{
decrypt(.data, ...)
}
\arguments{
\item{.data}{Dataframe containing crypted variable}

\item{...}{Columns to be decrypted}
}
\description{
Given a maskr crypted dataframe and a column variable, returns a decrypted
data frame.
}
\examples{
data(mtcars)
mtcars_encrypted <- encrypt(mtcars, cyl)
mtcars_decrypted <- decrypt(mtcars_encrypted, cyl)
identical(mtcars, mtcars_decrypted)

}
