% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt one column from a crypted dataframe using a dictionnary}
\usage{
decrypt(.data, .var, dictionnary)
}
\arguments{
\item{.data}{Dataframe containing crypted variable}

\item{.var}{Column to be decrypted}

\item{.dictionnary}{Dictionnary to be used to decrypt the crypted variable}
}
\description{
Given a maskr crypted dataframe and a dictionnary, returns a decrypted data
frame.
}

