% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt multiple columns from a dataframe}
\usage{
encrypt(.data, ...)
}
\arguments{
\item{.data}{Dataframe containing the variables}

\item{...}{Columns to be encrypted}
}
\description{
Given a dataframe and some column variables, returns a crypted data frame.
}
\examples{
data(mtcars)
mtcars_encrypted <- encrypt(mtcars, cyl)

}
