% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt one column from a  dataframe using a dictionary}
\usage{
encrypt(.data, .var)
}
\arguments{
\item{.data}{Dataframe containing the variable}

\item{.var}{Column to be crypted}
}
\description{
Given a maskr dataframe and a dictionary, returns a crypted data
frame.
}
