% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt one column from a dataframe}
\usage{
encrypt(.data, .var)
}
\arguments{
\item{.data}{Dataframe containing the variable}

\item{.var}{Column to be crypted}
}
\description{
Given a dataframe and a column variable, returns a crypted data frame.
}
\examples{
data(mtcars)
mtcars_encrypted <- encrypt(mtcars, cyl)

}
