% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Crypt one column from a  dataframe using a dictionnary}
\usage{
encrypt(.data, .var, dictionnary)
}
\arguments{
\item{.data}{Dataframe containing the variable}

\item{.var}{Column to be crypted}

\item{.dictionnary}{Dictionnary to be used to crypt the variable}
}
\description{
Given a maskr dataframe and a dictionnary, returns a crypted data
frame.
}

