% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt multiple columns from a dataframe}
\usage{
encrypt(.data, .dic, .randomize = TRUE)
}
\arguments{
\item{.data}{Dataframe containing the variables}

\item{.dic}{Dictionary containing the cryptograms}

\item{.randomize}{Randomize row order from \code{.data}. Defaults to \code{.randomize = TRUE}.}
}
\description{
Given a dataframe and some column variables, returns a crypted data frame.
}
\examples{
data(mtcars)
dic <- dictionary(mtcars, cyl, vs)
mtcars_encrypted <- encrypt(mtcars, dic)

}
