% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{MaskedArray}
\alias{MaskedArray}
\title{MaskedArray constructor function}
\usage{
MaskedArray(value, masks = default_mask(value))
}
\arguments{
\item{value}{An array-like object (vector, matrix, array, data.frame etc) to
mask}

\item{masks}{A list of vectors to use as the initial mask, one for each
dimension of \code{value}. \code{mask} and \code{masks} cannot both be specified. The
requirements for the \code{mask} parameter also apply to each constituent vector}

\item{mask}{A vector to use as the initial mask. It must be a vector that can
be used as a vector, so a character vector can only be used if \code{value} is
named. This will only work if \code{value} is a 1 dimensional type such as a
vector or 1D array.}
}
\description{
MaskedArray constructor function
}
